requirements.txt
flask==3.0.0
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.2
nltk==3.8.1
reportlab==4.0.7
requests==2.31.0
python-dotenv==1.0.0
xgboost==2.0.3
joblib==1.3.2
